requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2022.2.1=haa95532_0
certifi=2021.10.8=py38haa95532_2
click=8.0.4=py38haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=10.2.89=h74a9793_1
dataclasses=0.8=pyh6d0b6a4_7
flask=2.0.2=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
intel-openmp=2021.4.0=haa95532_3556
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.0.1=py38h2bbff1b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
numpy=1.21.5=py38ha4e8547_0
numpy-base=1.21.5=py38hc2deb75_0
openssl=1.1.1m=h2bbff1b_0
pillow=9.0.1=py38hdc2b20a_0
pip=21.2.2=py38haa95532_0
python=3.8.12=h6244533_0
pytorch=1.10.2=py3.8_cuda10.2_cudnn7_0
pytorch-mutex=1.0=cuda
setuptools=58.0.4=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=h2bbff1b_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.10.2=py38_cu102
torchvision=0.11.3=py38_cu102
tqdm=4.62.3=pyhd3eb1b0_1
typing_extensions=3.10.0.2=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0
